% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_MedMod.R
\name{MedMod}
\alias{MedMod}
\title{Mediated moderation for multiple mediators}
\usage{
MedMod(
  data_in,
  Yname,
  M1names,
  M2names,
  ttname,
  Rname,
  Cnames,
  learners = c("SL.glm"),
  num_folds = 1,
  Yfamily = NULL
)
}
\arguments{
\item{data_in}{a \code{data.frame} containing variables involved in the analysis.}

\item{Yname}{a character string of the name of the column in "data_in" that correspond to the outcome variable (Y).}

\item{M1names}{a character vector of the names of the columns in "data_in" that correspond to the mediator M1. The definition for the mediated moderation via mediator M1 (i.e., MedMod_M1) considers shifting the levels of M1, while fixing the levels of mediator M2 at its levels among the reference subgroup.}

\item{M2names}{a character vector of the names of the columns in "data_in" that correspond to the mediator M2. The definition for the mediated moderation via mediator M2 (i.e., MedMod_M2) considers shifting the levels of M2, while fixing the levels of mediator M1 at its levels among the focal subgroup.}

\item{ttname}{a character string of the name of the column in "data_in" that correspond to the treatment variable (T). The treatment variable should be dummy coded (e.g., 1 for the intervention condition, 0 for the control condition).}

\item{Rname}{a character string of the name of the column in "data_in" that correspond to the moderator subgroup status (R). The subgroup status should be dummy coded, with 1 for the focal subgroup and 0 for the reference subgroup.}

\item{Cnames}{a character vector of the names of the columns in "data_in" that correspond to pretreatment (i.e., baseline) covariates (C).}

\item{learners}{a character vector specifying the methods for estimating the nuisance models. Currently supported options include the list of methods included in the "SuperLearner" package (<https://cran.r-project.org/package=SuperLearner>), which can be listed via function listWrappers() of the "SuperLearner" package.}

\item{num_folds}{the number of folds used for the cross-fitting procedure.}

\item{Yfamily}{a character specifying the link function for a generalized linear model for the outcome. Currently supported options include "gaussian" and "binomial".}
}
\value{
A data frame containing the results for the mediated moderation estimands and the expected potential outcomes in the estimands. The results include the estimates, standard error estimates, and 0.95 confidence intervals.

 The column "Estimand" in the returned data frame contains the names of the estimands. Specifically:
 "MedMod_M1": The mediated moderation via mediator M1.
 "MedMod_M2": The mediated moderation via mediator M2.
 "MedMod_mu": The mediated moderation due to mutual dependence of the mediators M1 and M2.
 "MedMod_jo": The mediated moderation via both mediators M1 and M2 jointly. "MedMod_jo" is the sum of "MedMod_M1", "MedMod_M2", and "MedMod_mu".
 "RemainMod": The remaining moderation (i.e., not via (M1, M2)).
 "TotMod": The total moderation in the treatment effect on outcome between subgroups, controlling for the pretreatment covariates. "TotMod" is the sum of "RemainMod" and "MedMod_jo".
}
\description{
Mediated moderation for multiple mediators
}
